FIX #2553 Fusion Driver is supported only on darwin
