+++ drivers/vmwarefusion/fusion_darwin_test.go
