+++ drivers/vmwarefusion/vmx_darwin.go
